Add files via upload
This project implements an attendance system using Radio-Frequency Identification (RFID) technology and NodeMCU. Each person is assigned an RFID tag. When a tag is scanned by the NodeMCU, the system verifies its correctness and triggers a buzzer. Simultaneously, the attendee's is marked along with  name, current date and time. This data is updated on  Google Sheet, providing an automated and efficient way to track attendance in real-time.